Move metadata store to own module
