+++ crates/agent_ui/src/sidebar.rs
@@ -1,29 +1,21 @@
+use crate::thread_metadata_store::{ThreadMetadata, ThreadMetadataStore};
 use crate::threads_archive_view::{ThreadsArchiveView, ThreadsArchiveViewEvent};
 use crate::{Agent, AgentPanel, AgentPanelEvent, NewThread, RemoveSelectedThread};
 use acp_thread::ThreadStatus;
 use action_log::DiffStats;
-use agent::{ThreadStore, ZED_AGENT_ID};
-use agent_client_protocol::{self as acp, SessionId};
+use agent::ThreadStore;
+use agent_client_protocol::{self as acp};
 use agent_settings::AgentSettings;
-use anyhow::Result;
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 use db::kvp::KEY_VALUE_STORE;
-use db::{
-    sqlez::{
-        bindable::Column, domain::Domain, statement::Statement,
-        thread_safe_connection::ThreadSafeConnection,
-    },
-    sqlez_macros::sql,
-};
 use editor::Editor;
 use feature_flags::{AgentV2FeatureFlag, FeatureFlagViewExt as _};
 use gpui::{
-    Action as _, AnyElement, App, Context, Entity, FocusHandle, Focusable, Global, ListState,
-    Pixels, Render, SharedString, Subscription, Task, WeakEntity, Window, actions, list,
-    prelude::*, px,
+    Action as _, AnyElement, App, Context, Entity, FocusHandle, Focusable, ListState, Pixels,
+    Render, SharedString, WeakEntity, Window, actions, list, prelude::*, px,
 };
 use menu::{Cancel, Confirm, SelectFirst, SelectLast, SelectNext, SelectPrevious};
-use project::{AgentId, Event as ProjectEvent};
+use project::Event as ProjectEvent;
 use settings::Settings;
 use std::collections::{HashMap, HashSet};
 use std::mem;
@@ -40,298 +32,6 @@ use workspace::{
     MultiWorkspace, MultiWorkspaceEvent, ToggleWorkspaceSidebar, Workspace, multi_workspace_enabled,
 };
 use zed_actions::editor::{MoveDown, MoveUp};
-
-pub fn init(cx: &mut App) {
-    ThreadMetadataStore::init_global(cx);
-}
-
-struct GlobalThreadMetadataStore(Entity<ThreadMetadataStore>);
-impl Global for GlobalThreadMetadataStore {}
-
-pub struct ThreadMetadataStore {
-    db: ThreadMetadataDb,
-    session_subscriptions: HashMap<SessionId, Subscription>,
-}
-
-impl ThreadMetadataStore {
-    #[cfg(not(any(test, feature = "test-support")))]
-    pub fn init_global(cx: &mut App) {
-        let db = THREAD_METADATA_DB.clone();
-        let thread_store = cx.new(|cx| Self::new(db, cx));
-        cx.set_global(GlobalThreadMetadataStore(thread_store));
-    }
-
-    #[cfg(any(test, feature = "test-support"))]
-    pub fn init_global(cx: &mut App) {
-        let thread = std::thread::current();
-        let test_name = thread.name().unwrap_or("unknown_test");
-        let db_name = format!("THREAD_METADATA_DB_{}", test_name);
-        let db = smol::block_on(db::open_test_db::<ThreadMetadataDb>(&db_name));
-        let thread_store = cx.new(|cx| Self::new(ThreadMetadataDb(db), cx));
-        cx.set_global(GlobalThreadMetadataStore(thread_store));
-    }
-
-    pub fn global(cx: &App) -> Entity<Self> {
-        cx.global::<GlobalThreadMetadataStore>().0.clone()
-    }
-
-    pub fn list(&self, cx: &App) -> Task<Result<Vec<ThreadMetadata>>> {
-        let db = self.db.clone();
-        cx.background_spawn(async move {
-            let s = db.list()?;
-            Ok(s)
-        })
-    }
-
-    pub fn save(&mut self, metadata: ThreadMetadata, cx: &mut Context<Self>) -> Task<Result<()>> {
-        let db = self.db.clone();
-        cx.spawn(async move |this, cx| {
-            db.save(metadata).await?;
-            this.update(cx, |_this, cx| cx.notify())
-        })
-    }
-
-    pub fn delete(&mut self, session_id: SessionId, cx: &mut Context<Self>) -> Task<Result<()>> {
-        let db = self.db.clone();
-        cx.spawn(async move |this, cx| {
-            db.delete(session_id).await?;
-            this.update(cx, |_this, cx| cx.notify())
-        })
-    }
-
-    fn new(db: ThreadMetadataDb, cx: &mut Context<Self>) -> Self {
-        let weak_store = cx.weak_entity();
-
-        cx.observe_new::<acp_thread::AcpThread>(move |thread, _window, cx| {
-            let thread_entity = cx.entity();
-
-            let weak_store_2 = weak_store.clone();
-            cx.on_release(move |thread, cx| {
-                weak_store_2
-                    .update(cx, |store, _cx| {
-                        store.session_subscriptions.remove(thread.session_id());
-                    })
-                    .ok();
-            })
-            .detach();
-
-            weak_store
-                .update(cx, |this, cx| {
-                    let subscription = cx.subscribe(&thread_entity, Self::handle_thread_update);
-                    this.session_subscriptions
-                        .insert(thread.session_id().clone(), subscription);
-                })
-                .ok();
-        })
-        .detach();
-
-        Self {
-            db,
-            session_subscriptions: HashMap::new(),
-        }
-    }
-
-    fn handle_thread_update(
-        &mut self,
-        thread: Entity<acp_thread::AcpThread>,
-        event: &acp_thread::AcpThreadEvent,
-        cx: &mut Context<Self>,
-    ) {
-        match event {
-            acp_thread::AcpThreadEvent::NewEntry | acp_thread::AcpThreadEvent::TitleUpdated => {}
-            _ => return,
-        }
-
-        let thread_ref = thread.read(cx);
-        let session_id = thread_ref.session_id().clone();
-        let title = thread_ref.title();
-        let updated_at = Utc::now();
-
-        let agent_id = thread_ref.connection().agent_id();
-
-        let agent_id = if agent_id.as_ref() == ZED_AGENT_ID.as_ref() {
-            None
-        } else {
-            Some(agent_id)
-        };
-
-        let folder_paths = {
-            let project = thread_ref.project().read(cx);
-            let paths: Vec<Arc<Path>> = project
-                .visible_worktrees(cx)
-                .map(|worktree| worktree.read(cx).abs_path())
-                .collect();
-            PathList::new(&paths)
-        };
-
-        self.save(
-            ThreadMetadata {
-                session_id,
-                agent_id,
-                title,
-                updated_at,
-                created_at: None, //FIXME
-                folder_paths,
-            },
-            cx,
-        )
-        .detach_and_log_err(cx);
-    }
-}
-
-impl Global for ThreadMetadataStore {}
-
-/// Lightweight metadata for any thread (native or ACP), enough to populate
-/// the sidebar list and route to the correct load path when clicked.
-#[derive(Debug, Clone)]
-pub struct ThreadMetadata {
-    pub session_id: acp::SessionId,
-    /// `None` for native Zed threads, `Some("claude-code")` etc. for ACP agents.
-    pub agent_id: Option<AgentId>,
-    pub title: SharedString,
-    pub updated_at: DateTime<Utc>,
-    pub created_at: Option<DateTime<Utc>>,
-    pub folder_paths: PathList,
-}
-
-#[derive(Clone)]
-pub struct ThreadMetadataDb(ThreadSafeConnection);
-
-impl Domain for ThreadMetadataDb {
-    const NAME: &str = stringify!(ThreadMetadataDb);
-
-    const MIGRATIONS: &[&str] = &[sql!(
-        CREATE TABLE IF NOT EXISTS sidebar_threads(
-            session_id TEXT PRIMARY KEY,
-            agent_id TEXT,
-            title TEXT NOT NULL,
-            updated_at TEXT NOT NULL,
-            created_at TEXT,
-            folder_paths TEXT,
-            folder_paths_order TEXT
-        ) STRICT;
-    )];
-}
-
-db::static_connection!(THREAD_METADATA_DB, ThreadMetadataDb, []);
-
-impl ThreadMetadataDb {
-    /// Upsert metadata for a thread (native or ACP).
-    pub async fn save(&self, row: ThreadMetadata) -> anyhow::Result<()> {
-        let id = row.session_id.0.clone();
-        let agent_id = row.agent_id.as_ref().map(|id| id.0.to_string());
-        let title = row.title.to_string();
-        let updated_at = row.updated_at.to_rfc3339();
-        let created_at = row.created_at.map(|dt| dt.to_rfc3339());
-        let serialized = row.folder_paths.serialize();
-        let (folder_paths, folder_paths_order) = if row.folder_paths.is_empty() {
-            (None, None)
-        } else {
-            (Some(serialized.paths), Some(serialized.order))
-        };
-
-        self.write(move |conn| {
-            let sql = "INSERT INTO sidebar_threads(session_id, agent_id, title, updated_at, created_at, folder_paths, folder_paths_order) \
-                       VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7) \
-                       ON CONFLICT(session_id) DO UPDATE SET \
-                           agent_id = excluded.agent_id, \
-                           title = excluded.title, \
-                           updated_at = excluded.updated_at, \
-                           folder_paths = excluded.folder_paths, \
-                           folder_paths_order = excluded.folder_paths_order";
-            let mut stmt = Statement::prepare(conn, sql)?;
-            let mut i = stmt.bind(&id, 1)?;
-            i = stmt.bind(&agent_id, i)?;
-            i = stmt.bind(&title, i)?;
-            i = stmt.bind(&updated_at, i)?;
-            i = stmt.bind(&created_at, i)?;
-            i = stmt.bind(&folder_paths, i)?;
-            stmt.bind(&folder_paths_order, i)?;
-            stmt.exec()
-        })
-        .await
-    }
-
-    /// List all sidebar thread metadata, ordered by updated_at descending.
-    pub fn list(&self) -> anyhow::Result<Vec<ThreadMetadata>> {
-        self.select::<ThreadMetadata>(
-            "SELECT session_id, agent_id, title, updated_at, created_at, folder_paths, folder_paths_order \
-             FROM sidebar_threads \
-             ORDER BY updated_at DESC"
-        )?()
-    }
-
-    /// Look up a single thread by session ID.
-    pub fn get(&self, session_id: &acp::SessionId) -> anyhow::Result<Option<ThreadMetadata>> {
-        self.select_row_bound::<&str, ThreadMetadata>(
-            "SELECT session_id, agent_id, title, updated_at, created_at, folder_paths, folder_paths_order \
-             FROM sidebar_threads \
-             WHERE session_id = ?"
-        )?(session_id.0.as_ref())
-    }
-
-    /// Delete metadata for a single thread.
-    pub async fn delete(&self, session_id: acp::SessionId) -> anyhow::Result<()> {
-        let id = session_id.0.clone();
-        self.write(move |conn| {
-            let mut stmt =
-                Statement::prepare(conn, "DELETE FROM sidebar_threads WHERE session_id = ?")?;
-            stmt.bind(&id, 1)?;
-            stmt.exec()
-        })
-        .await
-    }
-
-    /// Delete all thread metadata.
-    pub async fn delete_all(&self) -> anyhow::Result<()> {
-        self.write(move |conn| {
-            let mut stmt = Statement::prepare(conn, "DELETE FROM sidebar_threads")?;
-            stmt.exec()
-        })
-        .await
-    }
-}
-
-impl Column for ThreadMetadata {
-    fn column(statement: &mut Statement, start_index: i32) -> anyhow::Result<(Self, i32)> {
-        let (id, next): (Arc<str>, i32) = Column::column(statement, start_index)?;
-        let (agent_id, next): (Option<String>, i32) = Column::column(statement, next)?;
-        let (title, next): (String, i32) = Column::column(statement, next)?;
-        let (updated_at_str, next): (String, i32) = Column::column(statement, next)?;
-        let (created_at_str, next): (Option<String>, i32) = Column::column(statement, next)?;
-        let (folder_paths_str, next): (Option<String>, i32) = Column::column(statement, next)?;
-        let (folder_paths_order_str, next): (Option<String>, i32) =
-            Column::column(statement, next)?;
-
-        let updated_at = DateTime::parse_from_rfc3339(&updated_at_str)?.with_timezone(&Utc);
-        let created_at = created_at_str
-            .as_deref()
-            .map(DateTime::parse_from_rfc3339)
-            .transpose()?
-            .map(|dt| dt.with_timezone(&Utc));
-
-        let folder_paths = folder_paths_str
-            .map(|paths| {
-                PathList::deserialize(&util::path_list::SerializedPathList {
-                    paths,
-                    order: folder_paths_order_str.unwrap_or_default(),
-                })
-            })
-            .unwrap_or_default();
-
-        Ok((
-            ThreadMetadata {
-                session_id: acp::SessionId::new(id),
-                agent_id: agent_id.map(|id| AgentId::new(id)),
-                title: title.into(),
-                updated_at,
-                created_at,
-                folder_paths,
-            },
-            next,
-        ))
-    }
-}
 
 actions!(
     agents_sidebar,
